v4-bridge/v4-outbox-publisher/requirements.txt
psycopg2-binary==2.9.9
kafka-python==2.0.2
prometheus-client==0.19.0
python-json-logger==2.0.7
